Fix bug in generation of invoices
The stream management changed in Pharo 10 and the way we were coupling Seaside and Artifact before cannot be used anymore. Until we get a better way to do it. I'm using a "hackish" way to do the export. This way requires to access the content of a ZnBufferedWriteStream. (This cannot work in all cases but it work in our case here, so let's put it as an extension and not mess with Zinc)